+++ Source/CTCoreMessage.m
@@ -586,7 +586,7 @@ char * etpan_encode_mime_header(char * phrase)
     NSString *msgContent = [[self rfc822] stringByReplacingOccurrencesOfString:@"\r\n" withString:@"\n"];
     NSData *msgContentAsData = [msgContent dataUsingEncoding:NSUTF8StringEncoding];
     NSMutableData *emlx = [NSMutableData data];
-    [emlx appendData:[[NSString stringWithFormat:@"%-10d\n", msgContentAsData.length] dataUsingEncoding:NSUTF8StringEncoding]];
+    [emlx appendData:[[NSString stringWithFormat:@"%-10lu\n", (unsigned long)msgContentAsData.length] dataUsingEncoding:NSUTF8StringEncoding]];
     [emlx appendData:msgContentAsData];
